refactor: reorganizar estructura de archivos y actualizar rutas de importación
