+++ frontend/src/components/ui/input.js
@@ -1,4 +1,3 @@
-
 export const Input = ({ type, value, onChange }) => {
     return <input type={type} value={value} onChange={onChange} />;
 };
